refactor: replace single ts_code with ts_codes list in API and frontend
- Add ts_codes field to ExecutionResult DAO model
- Pipeline writes ts_codes list in _finalize_result
- API list_backtest_results returns ts_codes array with names
- API get_prediction_stocks uses ts_codes list for single-stock fallback
- API get_trade_filter_options collects from both old and new fields
- Frontend shows single name or count for multiple stocks
- Fix LSTM predictor: astype(float) to handle None values in data

+++ backend/src/trade_alpha/models/lstm/predictor.py
@@ -22,10 +22,10 @@ class LSTMPredictor(BasePredictor):
             if len(stock) < normalization_window:
                 continue
             
-            data = stock[self.config.feature_fields].values
+            data = stock[self.config.feature_fields].values.astype(float)
             normalization_data = data[-normalization_window:]
             feed = data[-seq_len:]
-            
+
             mean = normalization_data.mean(axis=0)
             std = normalization_data.std(axis=0)
             std[std == 0] = 1.0
